requirements.txt
beautifulsoup4
requests
pandas
lxml
sqlalchemy
jupyter
pyodbc
python-dotenv
